+++ SB03-Findex-Team3/src/main/java/com/sprint/findex/dto/request/IndexInfoUpdateRequest.java
@@ -1,11 +1,12 @@
 package com.sprint.findex.dto.request;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 public record IndexInfoUpdateRequest(
     Integer employedItemsCount,
     LocalDate basePointInTime,
-    Integer baseIndex,
+    BigDecimal baseIndex,
     boolean favorite
 ) {
