+++ SB03-Findex-Team3/src/main/resources/schema.sql
@@ -10,7 +10,7 @@ CREATE TABLE index_info
     index_name           VARCHAR(240)     NOT NULL,
     employed_items_count INT              NOT NULL,
     base_point_in_time   DATE             NOT NULL,
-    base_index           INT              NOT NULL,
+    base_index           DECIMAL(10,2)              NOT NULL,
     source_type          VARCHAR(10)      NOT NULL,
     favorite             BOOLEAN          NOT NULL,
     UNIQUE (index_classification, index_name),
